Move + symlink vscode settings to ~/.config
This will work on Linux, and it seems slightly preferable to have the
real files live at a place without a space in the path and just symlink
to them from "~/Library/Application Support".

Also move ~/.config/vscode to ~/.config/Code just to keep everything in
one place, and update references to these paths as needed.