De-sudoify like_spotify_song
All I needed sudo for before was to get the Spotify icon, but instead I
can just copy the icon into a new Automator .app workflow, and use that
to run the applescript that displays the notification. No sudo and no
Dock icon!

As a side bonus, the workflow will display a spinning gear while running
which is a neat little indicator that it's doing something.